internal/gocdk: move static package into internal
