+++ internal/cmd/gocdk/demo.go
@@ -23,7 +23,7 @@ import (
 	"strings"
 
 	"github.com/spf13/cobra"
-	"gocloud.dev/internal/cmd/gocdk/static"
+	"gocloud.dev/internal/cmd/gocdk/internal/static"
 	"golang.org/x/xerrors"
 )
