+++ internal/cmd/gocdk/internal/static/_assets/biome/biome.json
