+++ internal/cmd/gocdk/internal/static/_assets/biome/main.tf
