+++ internal/cmd/gocdk/internal/static/_assets/biome/outputs.tf
