+++ internal/cmd/gocdk/internal/static/_assets/biome/secrets.auto.tfvars
