+++ internal/cmd/gocdk/internal/static/_assets/biome/variables.tf
