+++ internal/cmd/gocdk/internal/static/_assets/demo/blob/demo_blob.go
