+++ internal/cmd/gocdk/internal/static/_assets/demo/pubsub/demo_pubsub.go
