+++ internal/cmd/gocdk/internal/static/_assets/demo/runtimevar/demo_runtimevar.go
