+++ internal/cmd/gocdk/internal/static/_assets/demo/secrets/demo_secrets.go
