+++ internal/cmd/gocdk/internal/static/_assets/init/.gitignore
