+++ internal/cmd/gocdk/internal/static/_assets/init/.vscode/tasks.json
