+++ internal/cmd/gocdk/internal/static/_assets/init/README.md
