+++ internal/cmd/gocdk/internal/static/_assets/init/biomes/README.md
