+++ internal/cmd/gocdk/internal/static/_assets/init/biomes/dev/variables.tf
