+++ internal/cmd/gocdk/internal/static/static.go
