+++ internal/cmd/gocdk/internal/static/vfsdata.go
