+++ internal/testing/runchecks.sh
@@ -159,8 +159,8 @@ function cleanupstaticgo() {
 trap cleanupstaticgo EXIT
 pushd internal/cmd/gocdk/ &> /dev/null
 go run -mod=readonly generate_static.go -- "$tmpvfsdatago" &> /dev/null
-( diff -u static/vfsdata.go - < "$tmpvfsdatago" && echo "  OK" ) || {
-  echo "FAIL: gocdk compiled assets are out of date; please run go generate in internal/cmd/gocdk and commit the updated static/vfsdata.go" && result=1
+( diff -u internal/static/vfsdata.go - < "$tmpvfsdatago" && echo "  OK" ) || {
+  echo "FAIL: gocdk compiled assets are out of date; please run go generate in internal/cmd/gocdk and commit the updated internal/static/vfsdata.go" && result=1
 }
 popd &> /dev/null
