+++ internal/website/content/internal/cmd/gocdk/internal/static/_index.md
@@ -1,0 +1,4 @@
+---
+title: gocloud.dev/internal/cmd/gocdk/internal/static
+type: pkg
+---
