--- internal/website/content/internal/cmd/gocdk/static/_index.md
@@ -1,4 +1,0 @@
----
-title: gocloud.dev/internal/cmd/gocdk/static
-type: pkg
----
